Gender_Data_Lab/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.1
djangorestframework==3.15.2
fonttools==4.58.0
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
multidict==6.4.3
mysqlclient==2.2.7
numpy==2.2.5
openai==0.28.0
packaging==25.0
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.4.0
yarl==1.20.0
